Approx 12-16% faster encode/decode. 

From #17:
```ruby
hashids = Hashids.new("this is my salt")
hashids2 = Hashids2.new("this is my salt")

Benchmark.ips do |x|
  x.report("before encode") do
    hashids.encode(12345)
  end
  x.report("after encode") do
    hashids2.encode(12345)
  end
  x.compare!
end
Benchmark.ips do |x|
  x.report("before decode") do
    hashids.decode("NkK9")
  end
  x.report("after decode") do
    hashids2.decode("NkK9")
  end
  x.compare!
end
```

MRI 2.4.3 (similar result on 2.5.1):
```
Calculating -------------------------------------
       before encode     42.276k (± 2.3%) i/s -    214.604k in   5.079122s
        after encode     48.071k (± 2.2%) i/s -    242.216k in   5.041219s

Comparison:
        after encode:    48070.8 i/s
       before encode:    42276.5 i/s - 1.14x  slower

Warming up --------------------------------------
       before decode     2.042k i/100ms
        after decode     2.344k i/100ms
Calculating -------------------------------------
       before decode     20.767k (± 2.1%) i/s -    104.142k in   5.017175s
        after decode     23.797k (± 2.0%) i/s -    119.544k in   5.025694s

Comparison:
        after decode:    23796.6 i/s
       before decode:    20766.7 i/s - 1.15x  slower
```