@peterhellberg 2018-10-31
Thank you for your contribution, I will check it out later tonight and cut a new release of the gem.