@peterhellberg 2018-11-07
@tcrouch I've now cut 1.0.5 and pushed it to RubyGems: https://rubygems.org/gems/hashids/versions/1.0.5